Move forever repository to a data module
