+++ app/data/data-forever/build.gradle.kts
@@ -1,0 +1,14 @@
+plugins {
+  id("hedvig.android.ktlint")
+  id("hedvig.android.library")
+  alias(libs.plugins.squareSortDependencies)
+}
+
+dependencies {
+  implementation(libs.arrow.core)
+  implementation(libs.koin.core)
+  implementation(projects.apolloCore)
+  implementation(projects.apolloGiraffePublic)
+  implementation(projects.coreCommonPublic)
+  implementation(projects.languageCore)
+}
