+++ app/feature/feature-profile/src/main/kotlin/com/hedvig/android/feature/profile/payment/Campaign.kt
@@ -6,9 +6,6 @@ import giraffe.fragment.QuoteCartFragment
 import org.javamoney.moneta.Money
 import javax.money.MonetaryAmount
 
-@JvmInline
-value class CampaignCode(val code: String)
-
 data class Campaign(
   val displayValue: String?,
   val incentive: Incentive,
